// packages/sqlfu/src/vendor/typesql/types.ts
export type CrudQueryType = 'Select' | 'Insert' | 'Update' | 'Delete';
// ...
export type DatabaseClient = MySqlDialect | SQLiteDialect | LibSqlClient | BunDialect | D1Dialect | PgDielect;

export type TypeSqlDialect = DatabaseClient['type'];

export type SQLiteClient = SQLiteDialect['type'] | LibSqlClient['type'] | BunDialect['type'] | D1Dialect['type'];
export type MySqlClient = MySqlDialect['type'];

export type MySqlDialect = {
	type: 'mysql2';
	client: unknown;
	databaseVersion: string;
	schema: string;
	isVersion8: boolean;
};

export type SQLiteDialect = {
	type: 'better-sqlite3';
	client: DatabaseSync;
};

export type BunDialect = {
	type: 'bun:sqlite';
	client: DatabaseSync;
};

export type LibSqlClient = {
	type: 'libsql';
	client: DatabaseSync;
};

export type D1Dialect = {
	type: 'd1';
	client: DatabaseSync;
};

export type PgDielect = {
	type: 'pg',
	client: unknown
}

export type TypeSqlConfig = {
	databaseUri: string;
	sqlDir: string;
	outDir?: string;
	client: TypeSqlDialect;
	authToken?: string;
	attach?: string[];
	loadExtensions?: string[];
	includeCrudTables: string[];
	moduleExtension?: 'js' | 'ts';
	/**
	 * Optional list of schemas to include during introspection.
	 * Defaults to ['public'] if not specified.
	 */
	schemas?: string[];
};